armory_management/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
claudesync==0.7.1
click==8.1.8
click-completion==0.5.2
coloredlogs==15.0.1
configparser==7.2.0
contourpy==1.3.1
coverage==7.8.0
crontab==1.0.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
Django==5.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
dlib==19.24.6
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==25.2.10
fonttools==4.57.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pluggy==1.5.0
prettytable==3.16.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.1
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segno==1.6.6
setuptools==78.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sqlparse==0.5.3
sseclient-py==1.8.0
stringzilla==3.12.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wcwidth==0.2.13
